added validators

+++ controllers/orderController.js
@@ -3,10 +3,9 @@ const Product = require('../models/Product')
 const Seller = require('../models/Seller')
 const { StatusCodes } = require('http-status-codes')
 const CustomError = require('../errors/customError')
-const { findById, findByIdAndUpdate } = require('../models/Review')
 
 const getAllOrders = async(req, res) => {
-    const orders = await Order.find({})
+    const orders = await Order.find({user:req.user.userId})
     if(!orders){
         throw new CustomError('Order not found.')
     }
@@ -18,7 +17,7 @@ const getOrder = async (req, res) => {
         params: {id: orderId},
     } = req
 
-    const order = await Order.findOne({_id: orderId})
+    const order = await Order.findOne({_id: orderId, user:req.user.userId})
     if(!order){
         throw new CustomError('Order does not exist')
     }
@@ -31,7 +30,7 @@ const updateOrder = async (req, res) => {
     } = req
 
     const order = await Order.findByIdAndUpdate(
-        {_id: orderId},
+        {_id: orderId, user:req.user.userId},
         req.body,
         {new: true}
     )
@@ -44,15 +43,25 @@ const updateOrder = async (req, res) => {
 }
 
 const createOrder = async (req, res) => {
+
+    const product = await Product.findOne({_id: req.body.product}).lean()
+    const seller = await Seller.findOne({_id: product.seller}).lean()
+
+    /* if(product.inStock === 0){
+        throw new CustomError('Order failed, no product in stock')
+    } */
+
+    req.body.user = req.user.userId
     const order = await Order.create(req.body)
     if(!order){
         throw new CustomError('Could not create order')
     }
 
-    const product = await Product.findOne({_id: order.product})
-    const seller = await Seller.findOne({_id: product.seller}).lean()
-
+    product.hasSold = product.hasSold + 1
+    product.inStock = product.inStock - 1
     seller.productsSold = seller.productsSold + 1
+
+    await Product.findByIdAndUpdate({_id:product._id}, product)
     await Seller.findByIdAndUpdate({_id:seller._id}, seller)
 
     res.status(StatusCodes.CREATED).json( {order} )
@@ -63,7 +72,7 @@ const deleteOrder =  async(req, res) => {
         params: {id: orderId},
     } = req
 
-    const order = await Order.findByIdAndDelete({_id:orderId})
+    const order = await Order.findByIdAndDelete({_id:orderId, user:req.user.userId})
 
     if(!order){
         throw new CustomError('Order does not exist')
